@@ -1,6 +1,8 @@
 {
   "name": "metamodels/attribute_tablemulti",
   "description": "MetaModels attribute for dynamic table multi",
+  "license": "LGPL-3.0-or-later",
+  "type": "contao-bundle",
   "keywords": [
     "contao",
     "metamodels",
@@ -9,9 +11,6 @@
     "tablemulti",
     "nontranslatable"
   ],
-  "type": "contao-bundle",
-  "homepage": "https://now.metamodel.me/",
-  "license": "LGPL-3.0-or-later",
   "authors": [
     {
       "name": "Ronny Binder",
@@ -42,6 +41,7 @@
       "role": "Developer"
     }
   ],
+  "homepage": "https://now.metamodel.me/",
   "support": {
     "email": "mail@metamodel.me",
     "issues": "https://github.com/MetaModels/attribute_tablemulti/issues",
@@ -50,18 +50,19 @@
     "source": "https://github.com/MetaModels/attribute_tablemulti"
   },
   "require": {
-    "php": "^7.4",
+    "php": "^8.1",
     "ext-pdo": "*",
-    "contao-community-alliance/dc-general": "^2.2",
-    "contao/core-bundle": "^4.9.0, <4.13.0",
-    "metamodels/core": "^2.2",
-    "symfony/dependency-injection": "^4.4.6",
-    "symfony/http-kernel": "~4.4.13"
+    "contao-community-alliance/dc-general": "^2.3",
+    "contao/core-bundle": "^4.13.0 <5.0",
+    "doctrine/dbal": "^3.3.2",
+    "metamodels/core": "^2.3",
+    "symfony/dependency-injection": "^5.4",
+    "symfony/http-kernel": "^5.4"
   },
   "require-dev": {
     "contao/manager-plugin": "^2.1",
-    "menatwork/contao-multicolumnwizard-bundle": "^3.4",
-    "phpcq/all-tasks": "^1.2"
+    "menatwork/contao-multicolumnwizard-bundle": "^3.6",
+    "phpcq/runner-bootstrap": "^1.0@dev"
   },
   "autoload": {
     "psr-4": {
@@ -76,17 +77,18 @@
       "MetaModels\\AttributeTableMultiBundle\\Test\\": "tests/Test"
     }
   },
-  "extra": {
-    "contao-manager-plugin": "MetaModels\\AttributeTableMultiBundle\\ContaoManager\\Plugin",
-    "branch-alias": {
-      "dev-hotfix/2.2.2": "2.2.2-dev"
-    }
-  },
   "config": {
     "allow-plugins": {
       "contao-components/installer": false,
-      "contao/manager-plugin": false
+      "contao/manager-plugin": false,
+      "php-http/discovery": true
     },
     "sort-packages": true
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-feature/2.3.0": "2.3.x-dev"
+    },
+    "contao-manager-plugin": "MetaModels\\AttributeTableMultiBundle\\ContaoManager\\Plugin"
   }
 }
