@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_tablemulti.
  *
- * (c) 2012-2020 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -18,7 +18,7 @@
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2020 The MetaModels team.
+ * @copyright  2012-2024 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_tablemulti/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -41,21 +41,21 @@ class AttributeTypeFactory extends AbstractAttributeTypeFactory
      *
      * @var Connection
      */
-    private $connection;
+    private Connection $connection;
 
     /**
      * The string util.
      *
      * @var StringUtil|Adapter
      */
-    private $stringUtil;
+    private Adapter|StringUtil $stringUtil;
 
     /**
      * The Validator.
      *
      * @var Validator|Adapter
      */
-    private $validator;
+    private Adapter|Validator $validator;
 
     /**
      * {@inheritDoc}
@@ -69,7 +69,7 @@ class AttributeTypeFactory extends AbstractAttributeTypeFactory
         $this->validator  = $validator;
         $this->typeName   = 'tablemulti';
         $this->typeIcon   = 'bundles/metamodelsattributetablemulti/tablemulti.png';
-        $this->typeClass  = 'MetaModels\Attribute\TableMulti\TableMulti';
+        $this->typeClass  = TableMulti::class;
     }
 
     /**
