@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_tablemulti.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,15 +15,19 @@
  * @author     Andreas Dziemba <adziemba@web.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     David Maack <david.maack@arcor.de>
- * @copyright  2012-2019 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2012-2024 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_tablemulti/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
- 
+
 /**
  * MCW settings
  */
- 
-if (file_exists(TL_ROOT . '/system/config/module-multicolumnwizard.php')) {
-    include_once(TL_ROOT . '/system/config/module-multicolumnwizard.php');
+$rootDir = System::getContainer()->getParameter('kernel.project_dir');
+assert(\is_string($rootDir));
+
+/** @psalm-suppress UnresolvableInclude */
+if (\file_exists($rootDir  . '/system/config/module-multicolumnwizard.php')) {
+    include_once($rootDir  . '/system/config/module-multicolumnwizard.php');
 }
