@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_tablemulti.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -20,25 +20,23 @@
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     David Maack <david.maack@arcor.de>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2024 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_tablemulti/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
-$GLOBALS['TL_DCA']['tl_metamodel_rendersetting']['metapalettes']['tablemulti extends default'] = array
-(
-    '+advanced' => array('tablemulti_hide_tablehead'),
-);
+$GLOBALS['TL_DCA']['tl_metamodel_rendersetting']['metapalettes']['tablemulti extends default'] = [
+    '+advanced' => ['tablemulti_hide_tablehead'],
+];
 
 
-$GLOBALS['TL_DCA']['tl_metamodel_rendersetting']['fields']['tablemulti_hide_tablehead'] = array
-(
-    'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_rendersetting']['tablemulti_hide_tablehead'],
+$GLOBALS['TL_DCA']['tl_metamodel_rendersetting']['fields']['tablemulti_hide_tablehead'] = [
+    'label'       => 'tablemulti_hide_tablehead.label',
+    'description' => 'tablemulti_hide_tablehead.description',
     'exclude'   => true,
     'inputType' => 'checkbox',
-    'eval'      => array
-    (
+    'eval'      => [
         'tl_class' => 'clr w50'
-    ),
+    ],
     'sql'       => "varchar(1) NOT NULL default '0'",
-);
+];
