@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_tablemulti.
  *
- * (c) 2012-2022 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -16,7 +16,7 @@
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2022 The MetaModels team.
+ * @copyright  2012-2024 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_tablemulti/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -68,6 +68,12 @@ class PluginTest extends TestCase
         $bundleConfig = $bundles[0];
 
         self::assertEquals([MetaModelsCoreBundle::class], $bundleConfig->getLoadAfter());
-        self::assertEquals(['metamodelsattribute_multi', 'metamodelsattribute_tablemulti'], $bundleConfig->getReplace());
+        self::assertEquals(
+            [
+                'metamodelsattribute_multi',
+                'metamodelsattribute_tablemulti'
+            ],
+            $bundleConfig->getReplace()
+        );
     }
 }
