@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_tablemulti.
  *
- * (c) 2012-2022 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -17,7 +17,7 @@
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2022 The MetaModels team.
+ * @copyright  2012-2024 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_tablemulti/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -28,14 +28,14 @@ use MetaModels\AttributeTableMultiBundle\DependencyInjection\MetaModelsAttribute
 use MetaModels\AttributeTableMultiBundle\Attribute\AttributeTypeFactory;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
-use MetaModels\AttributeTableMultiBundle\Migration\ChangeTableNameMigration;
 
 /**
  * This test case test the extension.
  *
  * @covers \MetaModels\AttributeTableMultiBundle\DependencyInjection\MetaModelsAttributeRatingExtension
+ *
+ * @SuppressWarnings(PHPMD.LongClassName)
  */
 class MetaModelsAttributeTableMultiExtensionTest extends TestCase
 {
@@ -59,27 +59,13 @@ class MetaModelsAttributeTableMultiExtensionTest extends TestCase
      */
     public function testFactoryIsRegistered()
     {
-        $container = $this->getMockBuilder(ContainerBuilder::class)->getMock();
-
-        $container
-            ->method('setDefinition')
-            ->withConsecutive(
-                [
-                    'metamodels.attribute_tablemulti.factory',
-                    $this->callback(
-                        function ($value) {
-                            /** @var Definition $value */
-                            $this->assertInstanceOf(Definition::class, $value);
-                            $this->assertEquals(AttributeTypeFactory::class, $value->getClass());
-                            $this->assertCount(1, $value->getTag('metamodels.attribute_factory'));
-
-                            return true;
-                        }
-                    )
-                ]
-            );
+        $container = new ContainerBuilder();
 
         $extension = new MetaModelsAttributeTableMultiExtension();
         $extension->load([], $container);
+
+        self::assertTrue($container->hasDefinition('metamodels.attribute_tablemulti.factory'));
+        $definition = $container->getDefinition('metamodels.attribute_tablemulti.factory');
+        self::assertCount(1, $definition->getTag('metamodels.attribute_factory'));
     }
 }
